+++ onnxruntime/test/python/helper.py
@@ -1,0 +1,15 @@
+import os
+
+
+def get_name(name):
+    if os.path.exists(name):
+        return name
+    rel = os.path.join("testdata", name)
+    if os.path.exists(rel):
+        return rel
+    this = os.path.dirname(__file__)
+    data = os.path.join(this, "..", "testdata")
+    res = os.path.join(data, name)
+    if os.path.exists(res):
+        return res
+    raise FileNotFoundError(f"Unable to find '{name}' or '{rel}' or '{res}'")
